Fixed bean sequence, renamed package
