+++ client/spring-zeebe/src/main/java/io/camunda/zeebe/spring/client/metrics/DefaultNoopMetricsRecorder.java
@@ -1,4 +1,4 @@
-package io.camunda.zeebe.spring.client.connector;
+package io.camunda.zeebe.spring.client.metrics;
 
 /**
  * Default implementation for MetricsRecorder
