+++ client/spring-zeebe/src/main/java/io/camunda/zeebe/spring/client/metrics/MetricsRecorder.java
@@ -1,4 +1,4 @@
-package io.camunda.zeebe.spring.client.connector;
+package io.camunda.zeebe.spring.client.metrics;
 
 public interface MetricsRecorder {
