+++ connector-runtime/src/test/java/io/camunda/connector/runtime/inbound/WebhookControllerPlainJavaTests.java
@@ -47,8 +47,7 @@ import java.util.HashMap;
 import java.util.Map;
 import java.util.Set;
 
-import io.camunda.zeebe.spring.client.connector.DefaultNoopMetricsRecorder;
-import io.camunda.zeebe.spring.client.connector.MetricsRecorder;
+import io.camunda.zeebe.spring.client.metrics.MetricsRecorder;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
